refactor: FCM 관련 의존성 분리
